Enable hardhat plugin testing during CI
This diff's purpose is to enable running tests for hardhat in CI, and
also serve as an example of how a test launcher can be written in
bash.

At this moment, hardhat tests need you to manually start `nild`, as
they don't do it themselves. Fortunately, it's quite easy to do so in
a wrapper shell script, which I've implemented. The shell script will
make sure that `nild` is killed if the tests fail by setting a `trap`
for common signals (such as SIGINT or EXIT), so that you don't have to
worry about hanging background processes. This is quite standard for
writing automation shell scripts.